# Improve buffer allocation and copies for SiPixelDigisCUDA