--- a/CUDADataFormats/SiPixelCluster/src/SiPixelClustersCUDA.cc
+++ b/CUDADataFormats/SiPixelCluster/src/SiPixelClustersCUDA.cc
@@ -8,12 +8,12 @@ SiPixelClustersCUDA::SiPixelClustersCUDA(size_t maxModules, cudaStream_t stream)
       clusInModule_d(cms::cuda::make_device_unique<uint32_t[]>(maxModules, stream)),
       moduleId_d(cms::cuda::make_device_unique<uint32_t[]>(maxModules, stream)),
       clusModuleStart_d(cms::cuda::make_device_unique<uint32_t[]>(maxModules + 1, stream)) {
-  auto view = cms::cuda::make_host_unique<DeviceConstView>(stream);
+  auto view = cms::cuda::make_host_unique<SiPixelClustersCUDASOAView>(stream);
   view->moduleStart_ = moduleStart_d.get();
   view->clusInModule_ = clusInModule_d.get();
   view->moduleId_ = moduleId_d.get();
   view->clusModuleStart_ = clusModuleStart_d.get();
 
-  view_d = cms::cuda::make_device_unique<DeviceConstView>(stream);
+  view_d = cms::cuda::make_device_unique<SiPixelClustersCUDASOAView>(stream);
   cms::cuda::copyAsync(view_d, view, stream);
 }