--- a/CUDADataFormats/SiPixelDigi/interface/SiPixelDigiErrorsCUDA.h
+++ b/CUDADataFormats/SiPixelDigi/interface/SiPixelDigiErrorsCUDA.h
@@ -31,14 +31,14 @@ public:
   HostDataError dataErrorToHostAsync(cudaStream_t stream) const;
 
   void copyErrorToHostAsync(cudaStream_t stream);
-  int nErrors() const { return nErrors_; }
+  int nErrorWords() const { return nErrorWords_; }
 
 private:
   cms::cuda::device::unique_ptr<SiPixelErrorCompact[]> data_d;
   cms::cuda::device::unique_ptr<SiPixelErrorCompactVector> error_d;
   cms::cuda::host::unique_ptr<SiPixelErrorCompactVector> error_h;
   SiPixelFormatterErrors formatterErrors_h;
-  int nErrors_ = 0;
+  int nErrorWords_ = 0;
 };
 
 #endif  // CUDADataFormats_SiPixelDigi_interface_SiPixelDigiErrorsCUDA_h