--- a/CUDADataFormats/SiPixelDigi/interface/SiPixelDigisCUDA.h
+++ b/CUDADataFormats/SiPixelDigi/interface/SiPixelDigisCUDA.h
@@ -34,18 +34,9 @@ public:
 
 private:
   // These are consumed by downstream device code
-
-  int n16 = 4;
-  int n32 = 3;
-
   cms::cuda::device::unique_ptr<uint16_t[]> m_store;
 
   cms::cuda::host::unique_ptr<SiPixelDigisCUDASOAView> m_view;  // "me" pointer
-
-  // These are for CPU output; should we (eventually) place them to a
-  // separate product?
-  // cms::cuda::device::unique_ptr<uint32_t[]> pdigi_d;     // packed digi (row, col, adc) of each pixel
-  // cms::cuda::device::unique_ptr<uint32_t[]> rawIdArr_d;  // DetId of each pixel
 
   uint32_t nModules_h = 0;
   uint32_t nDigis_h = 0;