--- a/CUDADataFormats/SiPixelDigi/interface/SiPixelDigisCUDA.h
+++ b/CUDADataFormats/SiPixelDigi/interface/SiPixelDigisCUDA.h
@@ -10,7 +10,7 @@
 
 class SiPixelDigisCUDA {
 public:
-  using m_store_type = uint16_t;
+  using StoreType = uint16_t;
   SiPixelDigisCUDA() = default;
   explicit SiPixelDigisCUDA(size_t maxFedWords, cudaStream_t stream);
   ~SiPixelDigisCUDA() = default;
@@ -28,16 +28,16 @@ public:
   uint32_t nModules() const { return nModules_h; }
   uint32_t nDigis() const { return nDigis_h; }
 
-  cms::cuda::host::unique_ptr<m_store_type[]> copyAllToHostAsync(cudaStream_t stream) const;
+  cms::cuda::host::unique_ptr<StoreType[]> copyAllToHostAsync(cudaStream_t stream) const;
 
   SiPixelDigisCUDASOAView view() { return m_view; }
   SiPixelDigisCUDASOAView const view() const { return m_view; }
 
 private:
   // These are consumed by downstream device code
-  cms::cuda::device::unique_ptr<m_store_type[]> m_store;
+  cms::cuda::device::unique_ptr<StoreType[]> m_store;
 
-  SiPixelDigisCUDASOAView m_view;  // "me" pointer
+  SiPixelDigisCUDASOAView m_view;
 
   uint32_t nModules_h = 0;
   uint32_t nDigis_h = 0;