… EventFilter/SiPixelRawToDigi/plugins/SiPixelDigiErrorsSoAFromCUDA.cc
… EventFilter/SiPixelRawToDigi/plugins/SiPixelDigisSoAFromCUDA.cc
… RecoLocalTracker/SiPixelClusterizer/plugins/SiPixelRawToClusterCUDA.cc
… RecoLocalTracker/SiPixelClusterizer/plugins/SiPixelRawToClusterGPUKernel.cu
… RecoLocalTracker/SiPixelRecHits/plugins/PixelRecHitGPUKernel.cu
… RecoLocalTracker/SiPixelRecHits/plugins/SiPixelRecHitSoAFromLegacy.cc
… RecoLocalTracker/SiPixelRecHits/plugins/gpuPixelRecHits.h