#### PR description:

This is a technical PR that merges memory buffers of `SiPixelDigisCUDA`, to have a single `SoA` buffer. 
Also, output collections are now copied to the `CPU` in a single call. 

This decreases memory allocation for each thread, the number of certain API calls and improves throughput (measurements carried out in `pixeltrack-standalone framework`).

Other details:
* release: `CMSSW_12_1_0_pre5`
* machine: `cmg-gpu1080`
* `export CUDA_VISIBLE_DEVICES=4`

**API calls diff:** (for **9** events)
(measurement done with `nvprof`, output formatted with `awk`, `sort`, `diff`)
|  API name  |       baseline       |  digiscudasoa_patch_2021_11_17  | diff | per event diff | per reco diff |
| :-------------: |  :------------------:  | :---------------------------------------------:| :---:| :-----------------:|:---------------: |
| cudaEventCreateWithFlags | 218 | 206| -12 | | | 
| cudaEventDestroy | 218 | 206| -12 | | |
| cudaEventQuery | 1112 | 1032 | -80 | | |
| cudaEventRecord | 1080 | 990 | -90 | | |
| cudaFree | 164 | 160| -4 | | -4|
| cudaFreeHost | 299 | 291 | -8 | | -8 |
| cudaGetDevice | 2654 | 2470 | -184 | | |
| cudaHostAlloc | 298 | 290 | -8 | | -8 |
| cudaMalloc | 148 | 144 | -4 | | -4|
| cudaMemcpyAsync| 490| 454| -36| -4 | |
|cudaSetDevice| 1382| 1315 | -67 | | |
| cudaStreamWaitEvent | 3 | 6 | +3 | | | 

*diff view*
```
-218 cudaEventCreateWithFlags
+206 cudaEventCreateWithFlags
-218 cudaEventDestroy
+206 cudaEventDestroy
-1112 cudaEventQuery
+1032 cudaEventQuery
-1080 cudaEventRecord
+990 cudaEventRecord
-164 cudaFree
+160 cudaFree
-299 cudaFreeHost
+291 cudaFreeHost
-2654 cudaGetDevice
+2470 cudaGetDevice
-298 cudaHostAlloc
+290 cudaHostAlloc
-148 cudaMalloc
+144 cudaMalloc
-490 cudaMemcpyAsync
+454 cudaMemcpyAsync
-1382 cudaSetDevice
+1315 cudaSetDevice
-3 cudaStreamWaitEvent
+6 cudaStreamWaitEvent
```

**Memory allocation per event thread:**
(measurement done with `nvidia-smi` while running `cmsRun`)
|  threads/streams  |     baseline       |  digiscudasoa_patch_2021_11_17  | difference |
|----------|:-------------:|:------:|:-------------: |
| 1 thread/stream | 611 MiB | 597 MiB | -14 MiB |
| 8 threads/streams |    867 MiB   |  751 MiB | -116 MiB |
| diff - 7 threads | 256 MiB | 154 MiB |  - |
| **per thread** | **36 MiB** | **22 MiB** | **-14MiB** |

These measurements show that the **decrease** in **per event** memory allocations **scales with the number of threads** and is approximately **39%**.

**Throughput**
(in `pixeltrack-standalone` framework [applying the same changes](https://github.com/czangela/pixeltrack-standalone/tree/cudadev_digis_patch_2021_11_18))
```
# before
$  for N in 1 2 3 4; do taskset -c 0-15,32-47 ./cudadev  --numberOfThreads 16 --maxEvents 10000; done
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 8.657476e+00 seconds, throughput 1155.07 events/s.
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 8.695556e+00 seconds, throughput 1150.01 events/s.
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 8.730811e+00 seconds, throughput 1145.37 events/s.
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 8.742207e+00 seconds, throughput 1143.88 events/s.

# after
$  for N in 1 2 3 4; do taskset -c 0-15,32-47 ./cudadev  --numberOfThreads 16 --maxEvents 10000; done
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 8.489208e+00 seconds, throughput 1177.97 events/s.
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 8.525742e+00 seconds, throughput 1172.92 events/s.
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 8.563654e+00 seconds, throughput 1167.73 events/s.
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 8.579456e+00 seconds, throughput 1165.58 events/s.
```
**Before average: 1148,5825 events/s**
**After average: 1171,05 events/s**
`1171,05 / 1148,5825 - 1 = 1,956% ~ 2% speedup`

#### PR validation:

The reconstruction remains the same from the physics point of view.

**Validation in pixeltrack-standalone framework:**
```
$ ./cudadev --histogram --validation --numberOfThreads 16 --maxEvents 10000
Found 1 devices
Processing 10000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 10000 events passed validation
 Average relative track difference 0.000857969 (all within tolerance)
 Average absolute vertex difference 0.0011 (all within tolerance)
Processed 10000 events in 7.489605e+01 seconds, throughput 133.518 events/s.
```

#### if this PR is a backport please specify the original PR and why you need to backport that PR:

kindly ping @tonydp03 @felicepantaleo @VinInn