@cmsbuild 2021-11-19
+code-checks

Logs: https://cmssdt.cern.ch/SDT/code-checks/cms-sw-PR-36176/26742

- This PR adds an extra 36KB to repository

- There are other open Pull requests which might conflict with changes you have proposed:
  - File  RecoLocalTracker/SiPixelRecHits/plugins/SiPixelRecHitSoAFromLegacy.cc  modified in PR(s): #35713