@cmsbuild 2021-11-19
A new Pull Request was created by  @czangela  for master.

It involves the following packages:

- CUDADataFormats/SiPixelCluster (**heterogeneous, reconstruction**)
- CUDADataFormats/SiPixelDigi (**heterogeneous, reconstruction**)
- EventFilter/SiPixelRawToDigi (**reconstruction**)
- RecoLocalTracker/SiPixelClusterizer (**reconstruction**)
- RecoLocalTracker/SiPixelRecHits (**reconstruction**)


@jpata, @cmsbuild, @fwyzard, @makortel, @slava77 can you please review it and eventually sign? Thanks.
@mtosi, @OzAmram, @felicepantaleo, @GiacomoSguazzoni, @JanFSchulte, @rovere, @VinInn, @Martin-Grunewald, @missirol, @dkotlins, @ferencek, @gpetruc, @mmusich, @threus, @tvami this is something you requested to watch as well.
@perrotta, @dpiparo, @qliphy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>