@VinInn 2021-11-19
btw:
are we protected against ndigi==0?