@makortel 2021-11-23
In the PR description should these two numbers

> **Before average: 1171,05 events/s**
> **After average: 1148,5825 events/s**

be the other way around?