@cmsbuild 2021-11-25
+code-checks

Logs: https://cmssdt.cern.ch/SDT/code-checks/cms-sw-PR-36176/26890

- This PR adds an extra 24KB to repository

- There are other open Pull requests which might conflict with changes you have proposed:
  - File  RecoLocalTracker/SiPixelClusterizer/plugins/SiPixelRawToClusterGPUKernel.cu  modified in PR(s): #36235
  - File  RecoLocalTracker/SiPixelRecHits/plugins/SiPixelRecHitSoAFromLegacy.cc  modified in PR(s): #36235, #35713
  - File  RecoLocalTracker/SiPixelRecHits/plugins/gpuPixelRecHits.h  modified in PR(s): #36235