@czangela 2021-11-25
> btw: are we protected against ndigi==0?

Thank you! I'm still working on this.