@slava77 2021-11-29
> > btw: are we protected against ndigi==0?
> 
> Thank you! I'm still working on this.

I just wanted to check if this update is still planned for this PR.
Please clarify.
Thank you.