@czangela 2021-12-01
> > > btw: are we protected against ndigi==0?
> > 
> > 
> > Thank you! I'm still working on this.
> 
> I just wanted to check if this update is still planned for this PR. Please clarify. Thank you.

Yes, it is still planned for this PR.